requirements.txt
flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.1.0
nltk==3.8.1
google-generativeai==0.3.1
python-dotenv==1.0.0
